added reference plot
